+++ eventful-sqlite/tests/HLint.hs
@@ -7,13 +7,14 @@ import Prelude (String, IO, null)
 
 arguments :: [String]
 arguments =
-    [ "src"
-    , "tests"
-    , "-i=Redundant do"
-    , "-i=Unused LANGUAGE pragma" -- This fails on DeriveGeneric
-    ]
+  [ "src"
+  , "tests"
+  , "-i=Redundant do"
+  , "-i=Unused LANGUAGE pragma" -- This fails on DeriveGeneric
+  , "-i=Use section"
+  ]
 
 main :: IO ()
 main = do
-    hints <- hlint arguments
-    if null hints then exitSuccess else exitFailure
+  hints <- hlint arguments
+  if null hints then exitSuccess else exitFailure