+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Added a `GloballyOrderedProjection` abstraction to make projecting from
   globally-ordered event streams simpler.
 * Added useful `projectionMapMaybe` function.
+* Added the `runEventStoreUsing` and `runGloballyOrderedEventStoreUsing`
+  functions for running event stores in different monads than the original.
 
 ## 0.1.2